+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Update call result outcome design in contact patient bottom sheet
 - Add support for select and download in overdue search screen
 - Add support for select and share in overdue search screen
+- Show no internet connection dialog when download/share button is clicked
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Changes